I started by copying the existing typescript source plugin.
The readme is partially updated

the package.json is partially updated

the index.js is partially updated, but I am stumped on line 49 on where
to go next

The test cases are so far untouched. I'm committing them as-is from the
copy mostly as placeholders to fail for now because corresponding tests
probably need to be developed, or at least considered.